requirements.txt
Django>=4.2
pandas>=2.0
reportlab>=4.0
scikit-learn>=1.2
joblib>=1.3
numpy>=1.24
celery>=5.3
openpyxl>=3.1  # For Excel file support
python-dateutil>=2.8
djangorestframework>=3.14
channels>=4.0
channels-redis>=4.1
